Implement a Slack app
This is a very rought draft of a Slack app that can receive message
events and post a summary of topics to discuss in a thread. The tests
are minimal and only cover expected "happy" paths; the code is all in
one module with little encapsulation. This is deliberate. This project
will serve as an example of how an app might be architected better, so
showing the process of building in architecture is the point.